+++ src/storage.config.js
@@ -1,3 +1,0 @@
-export const config = {
-    storageType: "local", // 默认存储类型
-}