+++ src/utils/Data.js
@@ -1,10 +1,12 @@
 import { _Store } from "./data/Store.js";
+import { _Set } from "./set/Set.js";
 import { _All } from "./data/All.js";
 import { _ObjectValue } from "./data/ObjectValue.js";
 import { _ObjectMany } from "./data/ObjectMany.js";
 
 export {
     _Store,
+    _Set,
     _All,
     _ObjectValue,
     _ObjectMany