+++ src/utils/Helper.js
@@ -1,0 +1,5 @@
+import { getStorageSize } from "./size/storageSize.js";
+
+export {
+    getStorageSize, // 计算 webStorage 中所有数据的总大小
+};