+++ src/utils/data/value/set.js
@@ -1,16 +1,23 @@
 import { _Key, _Value } from "../../Parameter.js";
 
 /**
- * 从存储中获取指定键的值
- * @function getValueFromStorage
- * - 本函数中已经验证了 key 和 value 的有效性，不需要再外围再次验证，多次验证可能存在性能问题
- * @param { Storage } storage - 存储对象
- * @param { string } key - 要获取的键名
- * @returns { * } - 存储的值，如果没有找到则返回null
+ * 将数据存储到 web Storage 中
+ * - 本函数中已经验证了 key 的有效性，不需要再外围再次验证，多次验证可能存在性能问题
+ * @param { Storage } storage 存储对象
+ * @param { string } key 要存储的键名
+ * @param { * } value 要存储的数据
  */
 export function setValueToStorage(storage, key, value) {
     const KEY_ = _Key(key),
         VALUE_ = _Value(value);
 
-    storage.setItem(KEY_, JSON.stringify(VALUE_));
+    let cache = [];
+    storage.setItem(KEY_, JSON.stringify(VALUE_, (key, value) => {
+        if (typeof value === "object" && value !== null) {
+            if (cache.indexOf(value) !== -1) return;
+            cache.push(value);
+        }
+        return value;
+    }));
+    cache = null
 }